+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [1.0.0-alpha.10] - 2018-11-25
+### Added
+- Add express-mongo-sanitize middleware
+
+### Changed
+- Type property in abilities becomes not mandatory when ability has not state
+- Data property is not allowed in ability handlers when ability type is not defined
+
 ## [1.0.0-alpha.9] - 2018-11-18
 ### Added
 - Send entity operations events to all registered plugins